requirements.txt
certifi==2019.3.9
Django==2.2
djangorestframework==3.9.2
numpy==1.16.2
pytz==2018.9
sqlparse==0.2.4
